Move `execute_rule` tests
